add nonmutating operations (#25)
Implemented with no mutation, intended not to rely on either
`setindex!` or on `similar`.

Implemented in a point-free style.

The intention is to support `SMatrix` from StaticArrays.jl with no
allocation. However, the functionality is currently broken for
`SMatrix`. Not sure if that is a StaticArrays.jl bug or not.

The intention is to support `Matrix` without going overboard with
allocation.

Introduced a new module, `IndexArrays`, perhaps it will become a new
package in the future.

Implementing the MutableArithmetics.jl interface is left for a future PR.

Fixes #18